+++ lib/redis.utils.ts
@@ -1,5 +1,5 @@
+import Redis, { RedisOptions } from 'ioredis';
 import { RedisModuleOptions } from './redis.interfaces';
-import Redis from 'ioredis';
 import {
   REDIS_MODULE_CONNECTION,
   REDIS_MODULE_CONNECTION_TOKEN,
@@ -15,10 +15,19 @@ export function getRedisConnectionToken(connection?: string): string {
 }
 
 export function createRedisConnection(options: RedisModuleOptions) {
-  const { config } = options;
-  if (config.url) {
-    return new Redis(config.url, config);
-  } else {
-    return new Redis(config);
+  const { type, options: commonOptions } = options;
+
+  switch (type) {
+    case 'cluster':
+      return new Redis.Cluster(options.nodes, commonOptions);
+    case 'single':
+      const { url, options: { port, host } } = options;
+      const connectionOptions: RedisOptions = { ...commonOptions, port, host };
+
+      return url ? new Redis(url, connectionOptions) : new Redis(connectionOptions);
+    default:
+      throw new Error('Invalid configuration');
   }
 }
+
+